ui: draw the first screen without the slicing engine, and pass WCAG AA
The home dashboard was paying for the entire slicer before it could paint.
`Slicer` booted the runtime from its constructor, which on the web build
downloads the ~750 kB WebAssembly engine and starts the slicer worker — three
quarters of the page's bytes, and once compiled about 2 s of blocked main
thread, for a screen that slices nothing. It now boots on idle, and on demand
from every path that reaches the runtime, so a model dropped before the idle
callback fires simply claims the same promise a moment early.

The "Running in your browser" notice moves with it. Raised from `App`'s
constructor it was the largest block of text on screen, so it *was* the page's
Largest Contentful Paint and the whole site measured as loading however long
that dialog took to appear. It now fires when a model lands on the plate, which
is both cheaper and the moment its advice means anything.

Accessibility was failing two audits, both of them real:

- Zoom was disabled page-wide, by `user-scalable=no` in the viewport meta and
  `touch-action: none` on `html`, to keep gestures away from the 3D view.
  Between them they took magnification away from every settings form and every
  block of prose. The gesture lock now sits on the viewer's `:host`, which is
  the only surface that actually claims a pinch.
- `--color-text-tertiary` sat under 4.5:1 — 4.49:1 on the dark rail and 3.07:1
  in the light theme. Around eighty places consume it, so it is fixed at the
  token, in a local override emitted after the shared theme. The accent-tinted
  card in the WASM notice needed its own step up, since a neutral grey lands at
  3.9:1 on amber.

Also: a meta description (the one SEO miss), the in-app logo as WebP (52 kB of
PNG became 14 kB), and the dead `fonts.googleapis.com` preconnect dropped — the
production build inlines that stylesheet, so nothing ever contacted the host.

Measured on the built web bundle, served with gzip, no browser extensions,
median of three Lighthouse 13 mobile runs:

  performance      0.40 → 0.95
  accessibility    0.86 → 1.00
  best-practices   0.96 → 1.00
  seo              0.90 → 1.00

  total blocking time   2,010 ms → 23 ms
  largest contentful     10.8 s  → 2.7 s
  speed index             4.5 s  → 2.1 s
  page weight            1.51 MB → 0.98 MB
