+++ scripts/gen-logo-assets.sh
@@ -1,9 +1,12 @@
 #!/usr/bin/env bash
 #
-# gen-splash-logo.sh — regenerate the boot splash's two logo assets.
+# gen-logo-assets.sh — regenerate every logo asset the app ships, from one master.
 #
-# The splash in `ui/src/index.html` paints before any stylesheet or bundle has
-# arrived, so it shows the logo in two stages:
+# Two surfaces draw the mascot, and both are derived here so neither can drift
+# from `logo_still@3x.png`.
+#
+# **The boot splash** in `ui/src/index.html` paints before any stylesheet or
+# bundle has arrived, so it shows the logo in two stages:
 #
 #   1. a ~700-byte placeholder embedded in the document as base64, which costs
 #      no request at all and is therefore on screen with the first paint;
@@ -16,15 +19,21 @@
 # also strictly faster than a progressive format, whose first pass still costs a
 # round trip while an inlined placeholder costs none.
 #
+# **The in-app logo** (`ui/src/app/components/logo/`) sits in the header on every
+# screen. It used to be served as PNG, where the 2x variant a phone actually
+# picks weighed 52 KB for a 168px image — more than the rest of the page's
+# images together. The same artwork as WebP is 14 KB.
+#
 # WebP carries no PNG fallback deliberately: it has been supported everywhere
 # since Safari 14 (2020), and this app needs WebAssembly and WebGL2 to do
-# anything at all, so no browser that can run it lacks WebP.
+# anything at all, so no browser that can run it lacks WebP. The PNGs stay in
+# `ui/public/` as the masters this script reads, not as anything the app serves.
 #
 # Requires `cwebp` (brew install webp).
 #
 # Usage:
-#   scripts/gen-splash-logo.sh          # regenerate both assets
-#   scripts/gen-splash-logo.sh --check  # verify they are current (for CI)
+#   scripts/gen-logo-assets.sh          # regenerate every asset
+#   scripts/gen-logo-assets.sh --check  # verify they are current (for CI)
 set -euo pipefail
 
 cd "$(git rev-parse --show-toplevel 2>/dev/null || echo .)"
@@ -33,7 +42,7 @@ readonly SOURCE="ui/public/logo_still@3x.png"
 readonly FULL_ASSET="ui/public/splash-logo.webp"
 readonly INDEX="ui/src/index.html"
 
-# Scan every argument rather than reading `$1`: `pnpm run splash-logo -- --check`
+# Scan every argument rather than reading `$1`: `pnpm run logo-assets -- --check`
 # forwards the `--` separator as the first argument, so a positional read sees
 # `--` and silently takes the write path instead of checking.
 MODE=""
@@ -43,7 +52,7 @@ for arg in "$@"; do
     --) ;;
     *)
       echo "error: unknown argument: $arg" >&2
-      echo "usage: scripts/gen-splash-logo.sh [--check]" >&2
+      echo "usage: scripts/gen-logo-assets.sh [--check]" >&2
       exit 2
       ;;
   esac
@@ -71,25 +80,44 @@ fi
 tmp="$(mktemp -d)"
 trap 'rm -rf "$tmp"' EXIT
 
+# Emit `$2` from the freshly built `$1`, or verify it matches under --check.
+emit() {
+  local built="$1" asset="$2"
+  if [[ "$MODE" == "--check" ]]; then
+    if ! cmp -s "$built" "$asset"; then
+      echo "error: $asset is stale — run scripts/gen-logo-assets.sh" >&2
+      exit 1
+    fi
+    echo "$asset is up to date"
+  else
+    cp "$built" "$asset"
+    echo "wrote $asset ($(wc -c <"$asset" | tr -d ' ') bytes)"
+  fi
+}
+
 # Full-resolution asset.
 sips -Z "$FULL_PX" "$SOURCE" --out "$tmp/full.png" >/dev/null
 cwebp -quiet -q "$FULL_QUALITY" "$tmp/full.png" -o "$tmp/full.webp"
+emit "$tmp/full.webp" "$FULL_ASSET"
+
+# In-app logo. One file per `srcset` descriptor, so a phone downloads the 2x it
+# needs and nothing larger. Re-encoded from the PNG of the same size rather than
+# resampled from the master, so each stays pixel-identical to the artwork it
+# replaces.
+for scale in "" "@2x" "@3x"; do
+  png="ui/public/logo_still${scale}.png"
+  if [[ ! -f "$png" ]]; then
+    echo "error: logo master not found at $png" >&2
+    exit 1
+  fi
+  cwebp -quiet -q "$FULL_QUALITY" "$png" -o "$tmp/logo${scale}.webp"
+  emit "$tmp/logo${scale}.webp" "ui/public/logo_still${scale}.webp"
+done
 
 # Inline placeholder.
 sips -Z "$PLACEHOLDER_PX" "$SOURCE" --out "$tmp/lqip.png" >/dev/null
 cwebp -quiet -q "$PLACEHOLDER_QUALITY" -alpha_q 60 "$tmp/lqip.png" -o "$tmp/lqip.webp"
 data="data:image/webp;base64,$(base64 -i "$tmp/lqip.webp" | tr -d '\n')"
-
-if [[ "$MODE" == "--check" ]]; then
-  if ! cmp -s "$tmp/full.webp" "$FULL_ASSET"; then
-    echo "error: $FULL_ASSET is stale — run scripts/gen-splash-logo.sh" >&2
-    exit 1
-  fi
-  echo "$FULL_ASSET is up to date"
-else
-  cp "$tmp/full.webp" "$FULL_ASSET"
-  echo "wrote $FULL_ASSET ($(wc -c <"$FULL_ASSET" | tr -d ' ') bytes)"
-fi
 
 python3 - "$INDEX" "$data" "$MODE" <<'PY'
 import pathlib
