+++ ui/src/app/app.ts
@@ -5,7 +5,6 @@ import { NotificationCenter } from './components/notification-center/notificatio
 import { UpdateBanner } from './components/update-banner/update-banner';
 import { isTauriDesktop } from './runtime/domain/runtime-mode.util';
 import { AppVersion } from './services/app-version';
-import { WasmPerformanceNotice } from './services/wasm-performance-notice';
 import { DialogOutlet } from './shared/dialog/dialog-outlet';
 
 @Component({
@@ -17,16 +16,11 @@ import { DialogOutlet } from './shared/dialog/dialog-outlet';
 })
 export class App {
   private readonly appVersion = inject(AppVersion);
-  private readonly wasmPerfNotice = inject(WasmPerformanceNotice);
 
   constructor() {
     // Fire-and-forget: detect upgrades and surface "What's New" without
     // blocking startup. Failures are handled inside the service.
     void this.appVersion.checkForNewVersion();
-
-    // On the WASM web build only, remind the user once per session that
-    // in-browser slicing trades performance for zero install.
-    this.wasmPerfNotice.maybeShow();
 
     // Watch for a newer static deployment (Pages/web runtime) so a stale tab
     // gets a reload prompt even though there's no server to announce a version.
