+++ ui/src/app/components/logo/logo.html
@@ -1,7 +1,7 @@
 <img
   class="nexus-logo"
-  src="logo_still@2x.png"
-  srcset="logo_still.png 84w, logo_still@2x.png 168w, logo_still@3x.png 252w"
+  src="logo_still@2x.webp"
+  srcset="logo_still.webp 84w, logo_still@2x.webp 168w, logo_still@3x.webp 252w"
   sizes="86px"
   width="86"
   height="40"
