+++ ui/src/app/services/route-preload.ts
@@ -1,16 +1,7 @@
 import { Injectable } from '@angular/core';
 import type { PreloadingStrategy, Route } from '@angular/router';
 import { Observable, of, switchMap } from 'rxjs';
-
-/**
- * Give up waiting for a genuinely idle moment after this long and preload
- * anyway. A busy app (a slice running, a large model rendering) may never
- * report idle, and a user on that app is exactly the one about to navigate.
- */
-const IDLE_TIMEOUT_MS = 3_000;
-
-/** Fallback delay where `requestIdleCallback` is unavailable (Safari < 17). */
-const FALLBACK_DELAY_MS = 2_000;
+import { shouldConserveData, whenIdle } from './idle';
 
 /**
  * Fetch lazy route chunks in the background, once the app has nothing better
@@ -44,50 +35,4 @@ export class IdleRoutePreload implements PreloadingStrategy {
     }
     return whenIdle().pipe(switchMap(() => load()));
   }
-}
-
-/** Emits once the browser reports an idle moment (or the timeout expires). */
-function whenIdle(): Observable<void> {
-  return new Observable<void>((subscriber) => {
-    const done = () => {
-      subscriber.next();
-      subscriber.complete();
-    };
-
-    // Called as a member of `globalThis` rather than through a detached
-    // reference: `requestIdleCallback` is a Web IDL operation and throws
-    // "Illegal invocation" when its `this` is not the window.
-    if (typeof globalThis.requestIdleCallback !== 'function') {
-      const timer = setTimeout(done, FALLBACK_DELAY_MS);
-      return () => clearTimeout(timer);
-    }
-
-    const handle = globalThis.requestIdleCallback(done, { timeout: IDLE_TIMEOUT_MS });
-    return () => globalThis.cancelIdleCallback?.(handle);
-  });
-}
-
-/**
- * Whether the user has asked us not to spend bandwidth speculatively — either
- * explicitly (Data Saver) or implicitly by being on a slow connection.
- *
- * `navigator.connection` is Chromium-only, so this is a best-effort check: when
- * nothing is reported we preload, which is the right default for the desktop
- * and native builds where the assets are local anyway.
- */
-function shouldConserveData(): boolean {
-  const connection = (
-    navigator as Navigator & {
-      connection?: { saveData?: boolean; effectiveType?: string };
-    }
-  ).connection;
-
-  if (!connection) {
-    return false;
-  }
-  return (
-    connection.saveData === true ||
-    connection.effectiveType === '2g' ||
-    connection.effectiveType === 'slow-2g'
-  );
 }
